refactor: Organize component file structure
- move `ClassQuizList.jsx`, `Newsfeed.jsx`, `QuizForm.jsx` and `UserQuizList.jsx` to the `quizzes` repository under the `components` repository.
- move `Classroom.jsx` to `classroom repository` under the `components` repository.
- move `Chat.jsx` to `chat` repository under the `components` repository.
- change the import routes to match newly created file structure in `Dashboard.jsx` and `Quizzes.jsx` components.
